Account entity and create account EP created

+++ Backend/EzPay/src/main/java/com/ezpay/utils/dto/Account/AccountCreateDto.java
@@ -11,7 +11,7 @@ public record AccountCreateDto(
 
     @NotNull
     @Min(value = 1)
-    Long userId
+    Integer userId
 )
 {
